Initial implementation of scan io.Writer
It tends to be slightly slower than using normal scan and write.

When the payload is sufficiently large it starts to becode marginally
better, but not good enough to make the performance loss for small
values worth while. It's also not worth the additional code.

Preserving experiment on this branch for future reference.