Remove long TLS stress test
This was used to check that over 512 MB could be read over a TLS connection.
This previously could fail due to SSL renegotiation. But now pgx explicitly
disables renegotiation when connecting to the PostgreSQL server. Also, the Go
TLS library now supports limited renegotiation. And Amazon Redshift was the
only target that this mattered on, and it now supports disabling
renegotiation. So removing this long running and no longer needed test.