Add NewConnConfig callback to ConnPoolConfig
In CockroachDB test workloads, we want to have a pool of connections
against multiple database servers. With `sql.Database` and `lib/pq`,
we use our own driver which chooses one of multiple targets (in
round-robin fashion) on every `Open`.

In `pgx`, the same could be achieved using a custom `Dialer`, but this
would be pretty hacky:
 - we would need to encode the multiple server information into a
   single "address",
 - more importantly, it would break the cancellation logic which
   relies on creating a separate connection to the same server.

Instead, this commit adds a field `ConnPoolConfig.NewConnConfig` which
is a function that returns a `ConnConfig`. When this field is
specified, `ConnPoolConfig.ConnConfig` is ignored and this function is
used every time the pool creates a new connection. It's not great that
we have to keep both fields but it's necessary to maintain
compatibility.