Add MaxConnLifetime config setting to pool
in light of the issue we are experiencing with https://github.com/jackc/pgx/issues/845,
and the fact that we want to have the ability to recycle our connections,
this patch adds the MaxConnLifetime config value and its behaviour to v3 releases.

The work done in this PR is trying to mimic what has done for v4 track: https://github.com/jackc/pgx/commit/c604afba82679df5ca91fbd6da922fdd205d4310

The only difference is that this one doesn't have a default maxConnLifetime value defined,
which means that the current v3 usage out there shouldn't be impacted by this change.