fix prepared statement cache invalidation
This patch fixes jackc/pgx#841. To fix it we notice
when a statement failed because its cached plan is now
invalid and flush it from our cache. We don't try to get
fancy with retries or anything like that, just return the
error and allow the application to handle it.

The flushing of statements is deferred to the end of transactions
because all statements executed after an error in a transaction
will just result in errors. If we are not in a transaction, we
flush eagerly.

I had to make [some][1] [changes][2] to to the jackc/pgconn package as well
as this package, so I've retargeted this this repo on
ethanpailes/pgconn for the time being. This PR is not really in a
mergable state until those PRs merge and I rip out 2f31e98, but I
would say it is ready for review (just review the most recent commit).

One question I have is whether anything else needs to be done around
batch SQL, so guidance there would be appreciated.

Fixes #841

[1]: https://github.com/jackc/pgconn/pull/54
[2]: https://github.com/jackc/pgconn/pull/55