Call pgxpool AfterRelease inline, not in a goroutine
This is a partial revert of f572b336. This makes AfterRelease more useful since
it will always run before the Query methods return.

My specific use-case was adding events to a distributed trace. The problem was
that events were getting dropped because after the goroutine for AfterRelease
was consistently scheduled after request and corresponding trace ended. As an
example:

```go
func TestRacyAfterRelease(t *testing.T) {
	var connStr string
	ctx := context.Background()

	ctx, span := trace.OpenSpan(ctx, "test DB")
	cfg, _ := pgxpool.ParseConfig(connStr)
	cfg.BeforeConnect = func(ctx context.Context, cfg *pgx.ConnConfig) error {
		span.AddEvent("BeforeConnect")
		return nil
	}
	cfg.AfterRelease = func(conn *pgx.Conn) bool {
		span.AddEvent("AfterRelease")
		return true
	}
	pool, _ := pgxpool.ConnectConfig(ctx, cfg)

	dest := 0
	row := pool.QueryRow(ctx, `SELECT 1`)
	_ = row.Scan(&dest)

	// The "AfterRelease" event is probably not part of the trace since it
	// runs in a goroutine triggered by Scan.
	span.Close()
}
```

This commit runs AfterRelease inline instead of in the goroutine. If the caller
doesn't want to block on AfterRelease, they should create the goroutine, e.g.:

```go
	cfg.BeforeConnect = func(ctx context.Context, cfg *pgx.ConnConfig) error {
		go func() {
			span.AddEvent("BeforeConnect")
		}()
		return nil
	}
```