Adds CollectOneRowOK
CollectOneRowOK is a "comma OK idom" wrapper around CollectOneRow.

Compare CollectRows to CollectOneRow: when CollectRows returns no rows,
it returns an empty slice; but when CollectOneRow returns not rows,
it returns an error.

CollectOneRowOK provides a way to also return without
error when no rows are found. Whereas a caller of
CollectRows can check the len(returnedSlice) to determine if rows
were found, a caller of CollectOneRowOK can check the second return
value (usually named "ok") to see if rows were found.