Add QueuedQuery.Result()
This allows the user to have access to the pgconn.PgError early on
instead of after BatchResults.Close(). This allows for error parsing in
the QueuedQuery context. With the existing functions,
QueuedQuery.Query(), QueuedQuery.QueryRow() and QueuedQuery.Exec(), the
associated QueuedQuery has to be deduced.

Another reason to use QueuedQuery.Result() is that the other functions
result (potentially) in multiple kinds of error types: pgconn.PgError
and the error types returned from the functions passed in through
QueuedQuery.Query(), QueuedQuery.QueryRow() and QueuedQuery.Exec()