pgtype.TestTextCodeACLItem: Use built-in role instead of postgres
The role "postgres" is not guaranteed to exist. When creating a new
database with initdb, the default is the current user's name. Most
Linux distributions and many "production" Postgres installs will use
the name "postgres", but it is not the default. The
"pg_database_owner" role on the other hand is predefined and is
guaranteed to exist:
https://www.postgresql.org/docs/current/predefined-roles.html

This should make the tests easier to run for some users (like me!)

See initdb documentation for the --username flag for details:
https://www.postgresql.org/docs/current/app-initdb.html