pgtype: Change hstore tests to execute "create extension hstore"
The hstore extension can be loaded by any user with "create"
priviledges. The test already queries for the hstore OID, so we might
as well load the extension if it does not exist. This removes some
required setup for running the tests, so should make it a bit easier
for others to run them.