docs: Emphasise need to call rows.Err() after rows.Next() returns false
The Rows interface in pgx, like its ancestor in database/sql, is easy to
accidentally misuse in a way that can cause apps to misinterpret
database or connection errors as successful queries with empty or
truncated result-sets.

Update the docs to emphasise the need to call rows.Err() after
rows.Next() returns false, and direct users of the interface to the v5
API helpers that make writing correct code easier.

The docs on Conn.Query() already call this out, so only a small change
is needed to warn users and point them at the details on Query()

Per details in #1707